requirements.txt
python-dotenv 
pydub
google-generativeai
#audioop-lts
google-genai 
langchain
langchain_google_genai
langchain_experimental
langgraph
scikit-learn
pydantic
pandas
asyncio
groq
fastapi 
uvicorn
pyvirtualdisplay
matplotlib
seaborn
python-multipart
faiss-cpu
langchain_community
streamlit